# openapi.yaml
openapi: 3.0.0
info:
  title: Feature flag API
  version: 1.0.0
  description: This is API contract for feature flag service.
tags:
  - name: feature-flags
    description: Everything about feature flags
  - name: featureFlags-User
    description: Details about feature flags on Users.
  
paths:
  /feature-flags:
    get:
      tags:
        - feature-flags
      summary: Get all feature flags
      description: Returns a list of all feature flags in the system
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/ApiResponses/feature-flags'
        '404':
          description: Feature flags not found.
    post:
      tags:
        - feature-flags
      summary: Create a new feature flag
      description: Creates a new feature flag in the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/ApiRequests/feature-flag'
      responses:
        '201':
          description: Successful response
        '405': 
          description: Invalid input
  /feature-flags/{flagId}:
    get:
      tags:
        - feature-flags
      summary: Get a feature flag by ID
      description: Returns a specific feature flag based on ID
      parameters:
        - name: flagId
          in: path
          required: true
          description: ID of the user
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/ApiResponses/feature-flag'
        '404':
          description: feature flag not found
        '400':
          description: Invalid feature flag Id supplied
    patch:
      tags:
        - feature-flags
      summary: Update a feature flag by ID
      description: Updates a specific feature flag based on ID
      parameters:
        - name: flagId
          in: path
          required: true
          description: ID of the feature flag
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/ApiRequests/feature-flag-update'
      responses:
        '200':
          description: Successful response
        
          
  /users/{userId}/feature-flags/:
    get: 
      tags:
        - featureFlags-User
      summary: Lists all the feature flags for the user
      description: Shows the feature flag for a user.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema: 
                $ref: '#/ApiResponses/User-feature-flag-responses'
        '404':
          description: Any feature flags not enabled for the user

          
  /users/{userId}/feature-flag/{flagId}:
    get:
      tags:
        - featureFlags-User
      summary: Get feature flag for user
      description: Get if the flag is set for the user
      parameters:
        - name: userId
          in: path
          required: true
          description: Id of user
          schema: 
            type: string
        - name: flagId
          in: path
          required: true
          description: Id of the flag
          schema: 
            type: string
      responses:
        '200':
          description: Successful response
          content: 
            application/json:
              schema: 
                $ref: '#/ApiResponses/User-feature-flag-response'
        '400':
          description: Invalid input passed
          
    post:
      tags:
        - featureFlags-User
      summary: To enable feature flag for a user
      parameters:
        - name: userId
          in: path
          required: true
          description: Id of user
          schema: 
            type: 
              string
        - name: flagId
          in: path
          required: true
          description: Id of the flag
          schema: 
            type: string
      responses:
        '201':
          description: Created successful
        '405':
          description: Invalid input
    patch:
      tags:
        - featureFlags-User
      summary: To update feature flag for a user
      parameters:
        - name: userId
          in: path
          required: true
          description: Id of user
          schema: 
            type: 
              string
        - name: flagId
          in: path
          required: true
          description: Id of the flag
          schema: 
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/ApiRequests/featureflag-user-update'

      responses:
        '200':
          description: successful response
        '405':
          description: Invalid input
       
    

ApiResponses:
  feature-flag:
    type: object
    properties: 
      flagId:
        type: string
      flagName: 
        type: string
      flagDescription:
        type: string
      deactivated:
        type: boolean
        
  User-feature-flag-response:
    type: object
    properties:
      userId:
        type: string
      flagId:
        type: string
      enabled:
        type: boolean
  User-feature-flag-responses:
    type: array
    items: 
      type: object
      properties:
        userId:
          type: string
        flagId: 
          type: string
        enabled: 
          type: boolean
          
  feature-flags:        
    type: array
    items:
      type: object
      properties:
        flagId:
          type: string
        flagName:
          type: string
        flagDescription:
          type: string
        deactivated:
          type: boolean

ApiRequests:
  feature-flag:
    type: object
    properties:
      flagId:
        type: string
      flagName: 
        type: string
      flagDescription:
        type: string
  feature-flag-update:
    type: object
    properties: 
      flagId:
        type: string
      flagName:
        type: string
      status:
        type: string
      flagDescription:
        type: string
  
  featureflag-user-update:
    type: object
    properties:
      enabled:  
        type: boolean
    